Fixed a few issues with our tag_albums caching for tests.
