+++ backend/tests/unit/test_importer/conftest.py
@@ -1,7 +1,5 @@
-import atexit
 import logging
 import shutil
-import tempfile
 from pathlib import Path
 
 import pytest
@@ -68,74 +66,3 @@ def valid_data_for_album_path(path: str | Path) -> dict:
         }
     else:
         raise NotImplementedError(f"Unknown test album path {p=}")
-
-
-# ----------------- Monkeypath beets to use cached responses ----------------- #
-
-import hashlib
-import pickle
-
-from beets import autotag
-from beets.autotag import tag_album as _tag_album
-
-album_path: str
-lookup_cache_dir: Path
-
-
-def use_mock_tag_album(a_dir: str):
-    """Use a cached lookup for the tag_album function in beets
-    this allows to not make requests to the internet when testing
-    the importer.
-    """
-    global album_path, lookup_cache_dir
-    album_path = a_dir
-
-    # Create temp lookup cache directory
-    lookup_cache_dir = Path(tempfile.mkdtemp(prefix="beets_lookup_cache_"))
-    atexit.register(shutil.rmtree, lookup_cache_dir, ignore_errors=True)
-
-    autotag.tag_album = tag_album
-
-
-def tag_album(
-    items,
-    search_artist: str | None = None,
-    search_name: str | None = None,
-    search_ids: list[str] = [],
-):
-    global album_path
-    log.debug(f"Using monkey patched lookup {album_path=}")
-
-    # Compute items hash based on the items
-    m = hashlib.md5()
-    for item in items:
-        m.update(item.path)
-    if search_artist:
-        m.update(search_artist.encode("utf-8"))
-    if search_name:
-        m.update(search_name.encode("utf-8"))
-    for search_id in search_ids:
-        m.update(search_id.encode("utf-8"))
-    items_hash = m.hexdigest()[:8]
-
-    cache_file = lookup_cache_dir / f"lookup_{items_hash}.pickle"
-
-    if cache_file.exists():
-        log.debug(f"Using cached lookup from temp dir {cache_file}")
-        with open(cache_file, "rb") as f:
-            return pickle.load(f)
-    else:
-        # TODO: This pickle contains absolute paths to the files
-        # while undesired (no use in having them in the git repo) its for now the
-        # easiest way... and we hope music brainz does not change its data too often!
-        log.debug(f"Using default lookup {album_path=}")
-        res = _tag_album(items, search_artist, search_name, search_ids)
-
-        cache_file.parent.mkdir(parents=True, exist_ok=True)
-        with open(cache_file, "wb") as f:
-            pickle.dump(res, f)
-
-        return res
-
-
-autotag.tag_album = tag_album
